feat(maintner): adds sweeper utility
Because of issue transfers and deletes, it is possible for maintner to
have issues that remain in the corpus, despite being removed from the
GitHub repository.

This utility is designed to be used as a periodic Cron job which, for
each repository, queries the GitHub GraphQL API to retrieve the list of
issues for the repository, then, for each issue in the Maintner Corpus,
if that issue was not returned from GitHub, mark it as Tombstoned.

This introduces a new internal-only gRPC service to allow for a set of
issues to be marked as Tombstoned.