feat(maintner): remove CEL filtering
Filtering via the CEL was taking up an exorbitant amount of time when
making requests to maintner. Now using the `pull_request` and `closed`
fields in the proto.

NOTE: proto3 does not support nullable fields in messages so the option
of "pull_request = null" => both Pull Requests and Issues is not
easially supported. This was worked around by adding a `oneof` field,
which allows the sub members in it to be not specified.