feat(samplr): return NotFound status code
When the gRPC client attempts to dial to a service which does not exist,
it returns an "Unavailable" status code (which translates to a 503).

This is not entirely accurate to our callers as the service for that
REPOSITORY doesn't exist, but others might. Instead signal to the
calling clients that the resource is not found.