feat(maintner)!: gRPC API is only supported API (#24)
* feat(maintner)!: gRPC API is only supported API

Removes the "apiroutes" directory in maintnerd as the REST API is no longer
supported.

Adds gRPC health checking to maintner and maintner-rtr

The cloud build functionality now uses the API descriptor set for the
service definition as opposed to a swagger spec.

BREAKING_CHANGE: The "v0" API for maintner is no longer supported. The
"v1" API is still supported via HTTP => gRPC transcoding done by the
Extensible Service Proxy (ESP).

* fix(maintner): kubernetes manifest fixes

* fix(maintner-rtr): fix forwarding to port && registering health server

* fix(drghs): Makefile no longer needs v0

* feat(maintner): adds new message types

Adds support for Reviews and Comments on v1beta1

Adds Support for filtering via (depricated) flags in the
ListIssuesRequest message

ListIssuesRequest also contains flags to include reviews and comments

* fix(maintner) empty filters are true

* fix(maintner): fix gRPC error on nil unmarshal

* fix(maintner): fix bug where the correct repo was skipped

* fix(maintner): gRPC use Issue's Number as opposed to ID

* fix(maintner): rewrote API paths to be compatible

The v1 REST client defined the specification for "parents" as
"{owner}/{repo}" but samplr defines it as
"owners/{owner}/repositories/{repo}". Previous commits had made the
paths equivalent, but this broke backwards compatibility

* fix(maintner): fix issue ID comparison

* fix(maintner): message update

The v1 REST API's GetIssue RPC returned JSON of structure

{
  "Issue": {

  }
}

Where the Issue was a JSON object representing the actual Issue. This
commit updates the message contract to mimic that

* feat(maintner): add gRPC UnaryInterceptor for logs

* feat(maintnerd): add UnaryInterceptor for logs

* feat(maintner): remove CEL filtering

Filtering via the CEL was taking up an exorbitant amount of time when
making requests to maintner. Now using the `pull_request` and `closed`
fields in the proto.

NOTE: proto3 does not support nullable fields in messages so the option
of "pull_request = null" => both Pull Requests and Issues is not
easially supported. This was worked around by adding a `oneof` field,
which allows the sub members in it to be not specified.

* fix(maintner): fixed unit tests on filters

* chore(maintnerd): add more test to issue_service.go

* feat(maintner): add IssueServiceAdmin API

Adds new service IssueServiceAdmin. Has a single rpc:
UpdateTrackedRepos.

This was added to maintain parity with earlier behaviour.

maintner-rtr now implements IssueServiceAdminServer and forwards an
update RPC to maintnerd-sprvsr.

Note: this means that anyone with an API key can issue Update requests
and theoretically spam the service and cause a minor performance hit as
maintner-sprvsr has to hit the Kubernetes API, which in GKE has a rate
limit associated with it. This is guarded somewhat via a mutex in
sprvsr, but it is imperfect. Long term, we might want to reject
subsequent requests to Update if we are in the middle of an Update.

* fix(maintner): changed Update rpc to a POST

* fix(maintner): add URL to Issue proto response

* fix(endpoints): add magic-github-proxy back to deploy

* fix(maintner): add labels to makeIssuePB

* fix(maintner): add Commit, Repo, and ClosedAt to Issue

* feat(maintner): adds bloked, release_blocking and issue_type to Issue message

* fix(maintner): set PriorityUnknown as true by default