fix(samplrd): gRPC respects Context Cancels
The behaviour of errgroups is such that when the first routine returns
a non-nil error, the context associated with the group is Cancelled, the
Group waits for all routines to return, then returns the first non-nil
error.

Previously, the routine that was responsible for running the gRPC
service was not listening for cancellation signals from the Context, and
as such, would never return. This meant that if the Corpus's Sync
encoutnered a catastrophic error, the service would remain alive
indefinetly.

Now, if the Sync routine returns an error, the gRPC service will return
and the entire process with exit.