feat(maintner-swpr): several fixes and tests
Adds new function `buildLimiter` to build a rate.Limiter based on the
number of Issues and PullRequests for the given repository

Adds log statements for debugging

Adds a call to get the PullRequests from GitHub GraphQL API as, unlike
the standard REST API, Issues and PullReqeusts are treated as separate
types and are not returned in a joint call.

Adds constants to coordinate the ports for internal vs "v1" gRPC apis.

fixes bug with authenticating to GitHub by trimming the trailing "\n"
from the GITHUB_TOKEN env var.

fixes divide by 0 error in creating a rate.Limiter by using float
division as opposed to integer division