fix(maintner): no longer fail on ListIssues
In maintner-rtr when we were responding to a ListIssues RPC if any of
the deployments came back with an error, we would abandon the whole
request and return the error code (this included if we dialed to
a deployment). Now, log the error and continue on to the rest of the
errors.

This had a downstream effect on swpr (it could not do its work as the
list of repositories was always empty with an error).